+++ lib/prime_flutter.dart
@@ -13,6 +13,7 @@ export 'src/widgets/badge.dart';
 export 'src/widgets/bottom_sheet.dart';
 export 'src/widgets/button.dart';
 export 'src/widgets/divider.dart';
+export 'src/widgets/dropdown.dart';
 export 'src/widgets/expansion_list_item.dart';
 export 'src/widgets/filter_group.dart';
 export 'src/widgets/filter_pill.dart';